test(integration)#688: update the two-chamber RPC gate to the crypt topology
dungeon_two_chamber_test.go -> dungeon_crypt_test.go: StartEncounter's
default dungeon is now the crypt spec's 3-region entrance->corridor->boss
chain (2 connector doors), not the retired two-chamber generator (1
door), so the four gate facts this suite proves are generalized:

- chamberGoblins (region-ID-keyed) -> archetypeGoblins (RegionArchetype-
  keyed), since this package can't reference rpg-api's own unexported
  per-spec region ID constants and shouldn't need to -- the toolkit's
  fixed generic-role vocabulary (ArchetypeEntrance/ArchetypeBoss) is the
  right thing for a black-box RPC test to key off anyway.
- The pocket-scoping gate (fact 3) now sights an entrance-region goblin
  and asserts boss-region goblins stay out of initiative -- same shape as
  the old chamber-1/chamber-2 proof, renamed.
- The no-early-reveal gate (fact 4) now requires opening BOTH connector
  doors (closed doors block movement -- rpg-toolkit's own
  TestClosedDoor_BlocksMovement) before the sightline-forming move can
  reach the boss region, and checks EntityAppeared absence after each
  door individually AND after both, not just one.

lobby_start_then_move_test.go: doc-comment-only update noting
StartEncounter now calls InitDungeon (via a DungeonKey) instead of
InitTwoChamberRoom directly -- the entrance-edge-column property this
test's honesty check depends on is unchanged (InitDungeon's own
generateDungeonLayout places the entrance at region 0's far edge, same as
the generator it generalizes), and the test itself reads the spawn
position dynamically, so no logic changed.
